Ejercicio3

+++ Ejercicio3.java
@@ -10,7 +10,7 @@ public class Ejercicio3 {
     public static void main (String args []){
         Scanner sc = new Scanner(System.in);
         System.out.print("Dime la cantidad de pesetas: ");
-        double n = sc.nextInt();
+        double n = sc.nextDouble();
         double euros = 0.0060;
         
         System.out.print("La cantidad de "+ n + " pesetas en euros, son: " + n*euros);
